rename column headers in partner maps
