+++ data/streaming_services/partner_map.csv
@@ -1,4 +1,4 @@
-Partner Name,Company Name
+Company Name Source,Company Name
 8tracks,8tracks
 AWA,AWA
 Amazon AU BR MX Prime,Amazon
