+++ data/streaming_services/partner_map_simplified.csv
@@ -1,4 +1,4 @@
-Partner Name,Company Name
+Company Name Source,Company Name
 8tracks,Other
 Amazon AU BR MX Prime,Amazon
 Amazon Cloud Drive - DE,Amazon
