+++ src/main.py
@@ -3,7 +3,8 @@ import logging
 from pathlib import Path
 from bokeh.plotting import show
 
-import logic
+from data_processor import generate_reports, load_earnings_report
+import enums
 import plotting
 
 
@@ -12,8 +13,8 @@ logging.basicConfig(
     level=logging.INFO
 )
 
-allowed_distributors = [d.value for d in logic.Distributor]
-allowed_transactions = [t.value for t in logic.Transaction]
+allowed_distributors = [d.value for d in enums.Distributor]
+allowed_transactions = [t.value for t in enums.Transaction]
 
 
 @click.command()
@@ -26,24 +27,22 @@ def main(file_name, distributor, transactions) -> None:
     partner_map_path = Path('data/streaming_services/partner_map_simplified.csv')
     
     logging.info(f'Loading data from file: {source_data_path.name}...')
-    earnings_report = logic.load_earnings_report(source_data_path, distributor, partner_map_path)
+    earnings_report = load_earnings_report(source_data_path, distributor, partner_map_path)
 
     # TODO: separate reports into different functions
     # TODO: move the reports into a dataclass
     logging.info('Generating summary reports...')
-    summary_reports = logic.generate_reports(earnings_report, transactions, adjust_for_inflation=True)
-
-    print(earnings_report[earnings_report['Company Name'] == 'Unknown'])
+    summary_reports = generate_reports(earnings_report, transactions, adjust_for_inflation=True)
 
     # TODO: add transaction type(s) to plot title 
     logging.info('Generating interactive graph...')
 
-    transactions_str = ', '.join(transactions).title()
-    plot_1 = plotting.generate_bokeh_plot(summary_reports['rates'], title_text=f'{transactions_str} Transactions - Nominal')
-    plot_2 = plotting.generate_bokeh_plot(summary_reports['cpi_adjusted_rates'], title_text=f'{transactions_str} Transactions - Adjusted for inflation')
+    transactions_str = ', '.join(transactions).title() + 'Transactions - '
+    plot_1 = plotting.generate_bokeh_plot(summary_reports['rates'], title_text=f'{transactions_str} Nominal')
+    plot_2 = plotting.generate_bokeh_plot(summary_reports['cpi_adjusted_rates'], title_text=f'{transactions_str} Adjusted for inflation')
 
     show(plot_1)
     show(plot_2)
-        
+    
 if __name__ == '__main__':
     main()
